test: Consolidated tests for generation of netlists, SVG, XML, PCBs, schematics, etc.
